Feature-task4: *Refactor code
