--- src/com/lukashevich/shop/repository/impl/ProductRepositoryImpl.java
@@ -1,52 +1,0 @@
-package com.lukashevich.shop.repository.impl;
-
-import com.google.gson.Gson;
-import com.google.gson.GsonBuilder;
-import com.google.gson.reflect.TypeToken;
-import com.lukashevich.shop.model.Product;
-import com.lukashevich.shop.repository.ProductRepository;
-import com.lukashevich.shop.utils.FileUtils;
-
-import java.io.File;
-import java.io.FileWriter;
-import java.io.IOException;
-import java.lang.reflect.Type;
-import java.nio.file.Files;
-import java.nio.file.Path;
-import java.util.ArrayList;
-import java.util.List;
-
-public class ProductRepositoryImpl extends AbstractRepositoryImpl implements ProductRepository {
-    private final Gson gson;
-
-    public ProductRepositoryImpl() {
-        this.gson = new GsonBuilder().create();
-    }
-
-    @Override
-    public Product saveProduct(Product product) throws IOException {
-        File file = FileUtils.getOrCreateFile("ProductData.json");
-        List<Product> products = getAllProducts();
-        product.setId((long) products.size() + 1);
-        products.add(product);
-        try (FileWriter writer = new FileWriter(file, false)) {
-            writer.write(gson.toJson(products));
-            writer.flush();
-        }
-        return product;
-    }
-
-    @Override
-    public List<Product> getAllProducts() throws IOException {
-        File file = FileUtils.getOrCreateFile("ProductData.json");
-        if (file.length() == 0) {
-            return new ArrayList<>();
-        }
-        String productsJson = new String(Files.readAllBytes(Path.of(file.getPath())));
-        Type type = new TypeToken<ArrayList<Product>>() {
-        }.getType();
-        return gson.fromJson(productsJson, type);
-    }
-
-
-}
