+++ src/com/lukashevich/shop/repository/impl/ShopRepository.java
@@ -1,16 +1,15 @@
 package com.lukashevich.shop.repository.impl;
 
 import com.lukashevich.shop.model.Shop;
-import com.lukashevich.shop.repository.ShopRepository;
 
 
 import java.io.IOException;
 
 import java.util.List;
 
-public class ShopRepositoryImpl extends AbstractRepositoryImpl implements ShopRepository  {
+public class ShopRepository extends BaseRepository<Shop> implements com.lukashevich.shop.repository.ShopRepository {
 
-    public ShopRepositoryImpl() {
+    public ShopRepository() {
     }
 
     @Override
