+++ src/com/lukashevich/shop/service/impl/ProductServiceImpl.java
@@ -2,7 +2,6 @@ package com.lukashevich.shop.service.impl;
 
 import com.lukashevich.shop.model.Product;
 import com.lukashevich.shop.repository.ProductRepository;
-import com.lukashevich.shop.repository.impl.ProductRepositoryImpl;
 import com.lukashevich.shop.service.ProductService;
 
 import java.io.IOException;
