+++ src/com/lukashevich/shop/utils/FileUtils.java
@@ -5,8 +5,8 @@ import java.io.IOException;
 
 public class FileUtils {
 
-    public static File getOrCreateFile(String fileName) throws IOException {
-        File file = new File(fileName);
+    public static File getOrCreateFile(Class<?> tClass) throws IOException {
+        File file = new File(tClass.getSimpleName()+"Data.json");
         file.createNewFile();
         return file;
     }
